requirements.txt
alembic==1.15.2
aniso8601==10.0.0
apispec==6.8.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
exceptiongroup==1.2.2
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
flask-smorest==0.45.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mysql-connector-python==9.2.0
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.5
pytest-flask==1.3.0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
rpds-py==0.24.0
SQLAlchemy==2.0.40
swagger_ui_bundle==1.1.0
tomli==2.2.1
typing_extensions==4.13.0
webargs==8.6.0
Werkzeug==3.1.3
zipp==3.21.0
